refactor: consolidate React imports in bookmark window components
